Graphical and Mbc3 Bug Fixes
1. Rtc was not initializing on MBC3, added make() to struct
initialization.
2. window vs background is on a per pixel basis, not per scanline by
using the windowX Register. This is viewable with the Pokedex in Pokemon
Gold, where the left side of the screen uses background and the right
side uses window. Fixed by splitting background and window addresses per
scanline, then checking if window is used and if the current pixel >=
windowX. If you need a save to see behavior please let me know.